Iv 91 - Enhetsinformasjon

+++ src/main/resources/lib/cacheControll/index.es6
@@ -78,6 +78,7 @@ function wipeOnChange (path) {
         w(getPath(path, 'tavleliste'));
         w(getPath(path, 'tavleliste-relatert-innhold'));
         w(getPath(path, 'transport'));
+        w(getPath(path, 'office-information'));
         w(getPath(path, 'ekstraStorTabell'));
         if (path.indexOf('/driftsmeldinger/') !== -1) {
             w('driftsmelding-heading');
@@ -110,7 +111,7 @@ function activateEventListener () {
         localOnly: false,
         callback: function (event) {
             event.data.nodes.forEach(function (node) {
-                if (node.branch === 'master') {
+                if (node.branch === 'master' && node.repo === 'com.enonic.cms.default') {
                     wipeOnChange(node.path);
                     repo.query({
                         start: 0,
